Move derive_labelled_generic out
